requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic-settings
pydantic[email]
redis
chromadb
httpx
